tests/submission/W3C/domhtml/HTMLFormElement10.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>FORM3</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-form-element.html#dom-form-submit" rel="help">
<link rel="author" title="Curt Arnold">
</head>
<body>

<iframe id='iframe' height='200' width='550' srcdoc='
<form action="about:blank" id="form1">
<p>
<textarea cols="20" name="text1" rows="7"></textarea>
<input name="submit1" type="submit" value="Submit">
<input name="submit2" type="reset" value="Reset">
</p>
</form>'>
</iframe>



<script>
var domtest = async_test("HTMLFormElement.submit submits the form.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var doc;
      doc = window.document.getElementById('iframe').contentDocument;
      nodeList = doc.getElementsByTagName("form");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      testNode.submit();
      
    domtest.done();
}));

</script>
<p>HTMLFormElement.submit submits the form.</p>
<div id="log"></div>
</body></html>
